refactor : 불필요한 필드 삭제
