+++ src/main/java/app/bottlenote/review/dto/response/ReviewResponse.java
@@ -4,7 +4,6 @@ import app.bottlenote.review.domain.constant.ReviewDisplayStatus;
 import app.bottlenote.review.domain.constant.SizeType;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.List;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -27,21 +26,15 @@ public class ReviewResponse {
 	private final String userProfileImage;
 	private final Double rating;
 
-	private final String zipCode;
-	private final String address;
-	private final String detailAddress;
-
 	private final ReviewDisplayStatus status;
 
 	private final Boolean isMyReview;
 	private final Boolean isLikedByMe;
 	private final Boolean hasReplyByMe;
 
-	private List<String> reviewTastingTag;
-
 	@Builder
 	public ReviewResponse(Long reviewId, String reviewContent, BigDecimal price, SizeType sizeType, Long likeCount, Long replyCount, String reviewImageUrl, LocalDateTime createAt, Long userId, String nickName, String userProfileImage,
-		Double rating, String zipCode, String address, String detailAddress, ReviewDisplayStatus status, Boolean isMyReview, Boolean isLikedByMe, Boolean hasReplyByMe) {
+		Double rating, ReviewDisplayStatus status, Boolean isMyReview, Boolean isLikedByMe, Boolean hasReplyByMe) {
 		this.reviewId = reviewId;
 		this.reviewContent = reviewContent;
 		this.price = price;
@@ -54,16 +47,9 @@ public class ReviewResponse {
 		this.nickName = nickName;
 		this.userProfileImage = userProfileImage;
 		this.rating = rating;
-		this.zipCode = zipCode;
-		this.address = address;
-		this.detailAddress = detailAddress;
 		this.status = status;
 		this.isMyReview = isMyReview;
 		this.isLikedByMe = isLikedByMe;
 		this.hasReplyByMe = hasReplyByMe;
-	}
-
-	public void updateTastingTagList(List<String> reviewTastingTag) {
-		this.reviewTastingTag = reviewTastingTag;
 	}
 }
